requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.8.1
pillow==11.1.0
pytesseract==0.3.13
opencv-python-headless==4.11.0.86
pdf2image==1.17.0
streamlit==1.42.0
requests==2.32.3
pandas==2.2.3
mlflow==2.20.2
pytest==8.3.4
ruff==0.9.6
